+++ caQtDM_QtControls/src/imagewidget.cpp
@@ -166,8 +166,8 @@ void ImageWidget::paintEvent(QPaintEvent * event)
         pointsY.resize(YL.size());
         float Max[2];
         float Min[2];
-        Max[1] = SMALLEST;
-        Min[1] = BIGGEST;
+        Max[1] = (float)SMALLEST;
+        Min[1] = (float)BIGGEST;
 
         // get max and min from waveform
         for(int i=qMax(exposedRect.x(), 0); i< qMin(YL.size(),exposedRect.width() + exposedRect.x()) ; ++i) {
